Move Graph API to openclerk/api framework
Requires some changes to the GraphRenderer coffeescript since a standard
API response is now being returned

Issue #247
