+++ core/apis.json
@@ -1,3 +1,4 @@
 [
+  "\\Core\\Apis\\Graph",
   "\\Core\\Apis\\Rates"
 ]
